Move BodyContext and friends to their own file — groundwork for #87
Pure code motion, no behaviour change.  `BodyContext`, `VoiceAccumulator`,
`TupletState`, `applyTupletFactor` and its two fraction helpers move out of
`SemanticPass.swift` into `Semantic/BodyContext.swift`, taking 474 lines off a
1633-line file.

Nothing outside `SemanticPass.swift` referenced any of them, so the only edits
to the moved text are the two visibility changes that crossing a file boundary
requires: `BodyContext` and `TupletState` go from `private` to internal.
`TupletState` has to follow `BodyContext` because it is the type of a stored
property on it.  `applyTupletFactor`, `reduceFraction` and `gcdPrivate` stay
private — they are only reached from inside the new file.

Verified as motion rather than rewrite: the moved block diffs against the
original at exactly those two lines, and the change to `SemanticPass.swift` is
474 deletions with zero additions.  696 tests pass.
